Extract ShowError and ShowLoading
