+++ portal/src/ShowLoading.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ShowLoading: React.FC = function ShowLoading() {
+  return <div>Loading...</div>;
+};
+
+export default ShowLoading;
